Update release-solution-to-prod-with-inputs.yml
I think that this file has an inconsistency in the naming for the file between the 'convert-to-managed' and 'release-to-staging' jobs.  Removing "_managed" helped me.
The other two release-solution-to-prod sample-workflow files don't have the _managed suffix.

Also I had a timeout on the "Import solution as unmanaged to build env" step, which was resolved by adding "run-asynchronously: true" line.  Which I think is necessary for the "export async function main()" in https://github.com/microsoft/powerplatform-build-tools/blob/main/src/tasks/import-solution/import-solution-v2/index.ts